pythonanywhere_upload/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
--find-links https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu
transformers>=4.30.0
python-magic>=0.4.27
whitenoise>=6.5.0
django-storages>=1.14.0
boto3>=1.28.0
dj-database-url>=2.1.0 